Total re-structure coding of the objects: Created Booking manager instead of flight manager. Major algorithmic change, i.e Booking contains user and list of connecting flights.
A flight is a simple object from a to b cannot be changed. it is from the database.
A booking goes through the flight database, finds the linking cities between a to b and created a route( list of connecting flights) for the user booking.

New developer tasks will be assigned in the next meeting, and in the log book, method left to implement is the search.

will also update the log.txt right away